+++ packages/atomic/src/components/common/interface/interface-controller.ts
@@ -8,7 +8,7 @@ import type {AnyBindings, AnyEngineType} from './bindings';
 import {i18nBackendOptions, i18nTranslationNamespace, init18n} from './i18n';
 import '@/src/components/common/atomic-aria-live/atomic-aria-live';
 
-export type InitializeEventHandler = (bindings: AnyBindings) => void;
+type InitializeEventHandler = (bindings: AnyBindings) => void;
 export type InitializeEvent = CustomEvent<InitializeEventHandler>;
 
 export interface BaseAtomicInterface<EngineType extends AnyEngineType> {