+++ packages/atomic/src/components/common/layout/item-layout-controller.ts
@@ -12,7 +12,7 @@ export interface ItemLayoutHost extends ReactiveControllerHost {
   shadowRoot?: ShadowRoot | null;
 }
 
-export interface LayoutDisplayConfig {
+interface LayoutDisplayConfig {
   display: ItemDisplayLayout;
   density: ItemDisplayDensity;
   imageSize: ItemDisplayImageSize;